Refactor: Large profile refactor
Profile pages now have a dedicated layout. This is accessed alongside partials that represent profile pages.

This allows Users & Groups to reuse many of the same views in a shared location.

There are some cases where there is a check between user or group, for instance the profile buttons are model-specific.

The style tag in the profile_layout allows for style overwrite on specific pages. eg:
The friend display page can use this to set profile-sidebar to #display: none so that the profile-feed takes up the full width of the profile