Add: List of journals on main page & profiles
Main page journals include all journals that are authored by
friends of the current user, excluding journals made to private groups
that the current_user is not a member of.

Profile journals show all journals directed to the profile owner.
By default, a post made by a user is directed to themself.

Journals made by user on another user (or group) profile will not show
on their own profile. example:
User A posts a journal to User B's page.
This post will show on:
User B's profile
the main feed of friends of user A or B

The post will not show on:
User A's profile

However I am considering a seperate profile tab for including these