Add: Notifications
Notifications are created by callbacks in respective models.
Notifications are displayed on the right sidebar.

Notifications include:
A Journal being made to your page
A Journal being made to the page of a group you own
A comment reply on your Journal or Comment
A like on your journal or comment
A friend request has been accepted

Incoming friend requests are not a notification and are listed seperate